Meme Generator/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
docx==0.2.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
pathlib==1.0.1
Pillow==8.3.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
Werkzeug==2.0.1
